src/retriever.py
"""
Retriever module for semantic document search using FAISS.

Provides functions to perform similarity-based lookups over embedded document vectors.
Integrates with FAISS for efficient vector search and returns relevant document matches.
"""

import faiss
import numpy as np


def search_documents(query: str, config: dict):
    # TODO: Load FAISS index and metadata
    # For now simulate with dummy results

    return [f"Match for '{query}' in file1.py", f"Match in utils.py"]
