requirements.txt

# Core ML and NLP dependencies
torch>=2.0.0
transformers>=4.35.0
sentence-transformers>=2.2.2
accelerate>=0.24.0
tokenizers>=0.15.0

# Vector storage and search
faiss-cpu>=1.7.4
chromadb>=0.4.15

# CLI framework and utilities
typer>=0.9.0
click>=8.1.0
rich>=13.6.0
pydantic>=2.5.0

# Configuration and data handling
pyyaml>=6.0.1
toml>=0.10.2
python-dotenv>=1.0.0

# Git integration for diff analysis
gitpython>=3.1.40
pygments>=2.16.0

# File processing and I/O
pathlib2>=2.3.7
tqdm>=4.66.0
pandas>=2.1.0

# Quantization and optimization (for Phi-2)
bitsandbytes>=0.41.0
optimum>=1.14.0

# HTTP server (for optional API mode)
fastapi>=0.104.0
uvicorn>=0.24.0
httpx>=0.25.0

# Testing framework
pytest>=7.4.0
pytest-asyncio>=0.21.0
pytest-cov>=4.1.0

# Development tools
black>=23.9.0
isort>=5.12.0
flake8>=6.1.0
mypy>=1.6.0
